+++ client/components/AdminBoard.js
@@ -1,0 +1,9 @@
+import React, { useEffect, useState } from 'react'
+
+export const AdminBoard = () => {
+  return (
+    <div>
+      <h1>You've reached the Admin Board</h1>
+    </div>
+  )
+}
